requirements.txt
django-cors-headers==4.3.0
celery==5.3.0
Django==4.2.1
djangorestframework==3.14.0
python-dotenv==1.0.0
redis==4.5.5
requests==2.31.0
xmltodict==0.13.0
gunicorn==20.1.0
gevent==22.10.2
psycopg2==2.9.6
pycryptodome==3.18.0
PyJWT==2.7.0
django-redis==6.0.0